secure-rag-app/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-cloud-storage
google-cloud-pubsub
pydantic
chromadb
sentence-transformers
google-cloud-aiplatform
python-jose[cryptography]
python-magic
